move runner inputs module
