+++ runzi/runners/average_solutions.py
@@ -4,7 +4,7 @@ import logging
 from argparse import ArgumentParser
 from pathlib import Path
 
-from runzi.automation.runner_inputs import AverageSolutionsInput
+from runzi.runners.runner_inputs import AverageSolutionsInput
 from runzi.automation.scaling.local_config import API_KEY, API_URL, USE_API
 from runzi.automation.scaling.schedule_tasks import schedule_tasks
 from runzi.automation.scaling.task_utils import get_model_type
