+++ runzi/runners/coulomb_rupture_sets.py
@@ -9,7 +9,7 @@ from multiprocessing.dummy import Pool
 from pathlib import Path, PurePath
 from subprocess import check_call
 
-from runzi.automation.runner_inputs import CoulombRuptureSetsInput
+from runzi.runners.runner_inputs import CoulombRuptureSetsInput
 from runzi.automation.scaling import coulomb_rupture_set_builder_task
 from runzi.automation.scaling.local_config import (
     API_KEY,
